refactor(GreetingForm): move off GreetingForm from greeting component
